requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
absl-py=1.0.0=pyhd8ed1ab_0
aiohttp=3.7.4.post0=py38h7f8727e_2
antlr-python-runtime=4.8=pyhd8ed1ab_3
async-timeout=3.0.1=py_1000
attrs=21.4.0=pyhd8ed1ab_0
bio-embeddings-esm=0.2.1=pyhd8ed1ab_0
biopython=1.79=py38h497a2fe_0
blas=1.0=mkl
blinker=1.4=py_1
bottleneck=1.3.4=py38hce1f21e_0
brotlipy=0.7.0=py38h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
c-ares=1.18.1=h7f8727e_0
ca-certificates=2021.10.8=ha878542_0
cachetools=5.0.0=pyhd8ed1ab_0
certifi=2021.10.8=py38h578d9bd_1
cffi=1.15.0=py38hd667e15_1
chardet=4.0.0=py38h578d9bd_2
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=7.1.2=pyh9f0ad1d_0
colorama=0.4.4=pyh9f0ad1d_0
configparser=5.2.0=pyhd8ed1ab_0
cryptography=36.0.0=py38h9ce1e76_0
cudatoolkit=11.3.1=h2bc3f7f_2
dataclasses=0.8=pyhc8e2a94_3
dendropy=4.5.2=pyh3252c3a_0
docker-pycreds=0.4.0=py_0
faiss-cpu=1.7.2=py3.8_h2a577fa_0_cpu
ffmpeg=4.3=hf484d3e_0
freetype=2.11.0=h70c0345_0
fsspec=2022.2.0=pyhd8ed1ab_0
future=0.18.2=py38h578d9bd_4
giflib=5.2.1=h7b6447c_0
gitdb=4.0.9=pyhd8ed1ab_0
gitpython=3.1.27=pyhd8ed1ab_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
google-auth=2.6.2=pyh6c4a22f_0
google-auth-oauthlib=0.4.6=pyhd8ed1ab_0
grpcio=1.42.0=py38hce63b2e_0
hydra-core=1.1.1=pyhd8ed1ab_0
idna=3.3=pyhd3eb1b0_0
importlib-metadata=4.11.3=py38h578d9bd_0
importlib_metadata=4.11.3=hd8ed1ab_0
importlib_resources=5.6.0=pyhd8ed1ab_0
intel-openmp=2021.4.0=h06a4308_3561
jpeg=9d=h7f8727e_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libfaiss=1.7.2=h2bc3f7f_0_cpu
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgomp=9.3.0=h5101ec6_17
libiconv=1.15=h63c8f33_5
libidn2=2.3.2=h7f8727e_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.15.8=h780b84a_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuv=1.40.0=h7b6447c_0
libwebp=1.2.2=h55f646e_0
libwebp-base=1.2.2=h7f8727e_0
lz4-c=1.9.3=h295c915_1
markdown=3.3.6=pyhd8ed1ab_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
multidict=5.2.0=py38h7f8727e_2
ncurses=6.3=h7f8727e_2
nettle=3.7.3=hbbd107a_1
numexpr=2.8.1=py38h6abb31d_0
numpy=1.21.2=py38h20f2e39_0
numpy-base=1.21.2=py38h79a1101_0
oauthlib=3.2.0=pyhd8ed1ab_0
omegaconf=2.1.1=py38h578d9bd_1
openh264=2.1.1=h4ff587b_0
openssl=1.1.1n=h7f8727e_0
packaging=21.3=pyhd8ed1ab_0
pandas=1.3.1=py38h8c16a72_0
pandas-flavor=0.2.0=py_0
pathtools=0.1.2=py_1
pillow=9.0.1=py38h22f2fdc_0
pip=21.2.4=py38h06a4308_0
promise=2.3=py38h578d9bd_5
protobuf=3.15.8=py38h709712a_0
psutil=5.8.0=py38h27cfd23_1
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.21=pyhd3eb1b0_0
pydeprecate=0.3.2=pyhd8ed1ab_0
pyjwt=2.3.0=pyhd8ed1ab_1
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.7=pyhd8ed1ab_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.12=h12debd9_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.8=2_cp38
pytorch=1.11.0=py3.8_cuda11.3_cudnn8.2.0_0
pytorch-lightning=1.5.10=pyhd8ed1ab_0
pytorch-mutex=1.0=cuda
pytz=2021.3=pyhd3eb1b0_0
pyu2f=0.1.5=pyhd8ed1ab_0
pyyaml=5.4.1=py38h497a2fe_0
readline=8.1.2=h7f8727e_1
requests=2.27.1=pyhd3eb1b0_0
requests-oauthlib=1.3.1=pyhd8ed1ab_0
rsa=4.8=pyhd8ed1ab_0
sentry-sdk=1.5.8=pyhd8ed1ab_0
setproctitle=1.2.2=py38h497a2fe_0
setuptools=58.0.4=py38h06a4308_0
shortuuid=1.0.8=py38h578d9bd_0
six=1.16.0=pyhd3eb1b0_1
smmap=3.0.5=pyh44b312d_0
sqlite=3.38.0=hc218d9a_0
tensorboard=2.8.0=pyhd8ed1ab_1
tensorboard-data-server=0.6.0=py38h2b97feb_0
tensorboard-plugin-wit=1.8.1=pyhd8ed1ab_0
termcolor=1.1.0=py_2
tk=8.6.11=h1ccaba5_0
torchmetrics=0.7.3=pyhd8ed1ab_0
torchvision=0.12.0=py38_cu113
tqdm=4.63.1=pyhd8ed1ab_0
typing-extensions=4.1.1=hd3eb1b0_0
typing_extensions=4.1.1=pyh06a4308_0
urllib3=1.26.8=pyhd3eb1b0_0
wandb=0.12.11=pyhd8ed1ab_0
werkzeug=2.0.3=pyhd8ed1ab_1
wheel=0.37.1=pyhd3eb1b0_0
xarray=2022.3.0=pyhd8ed1ab_0
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h516909a_0
yarl=1.6.3=py38h27cfd23_0
yaspin=2.1.0=pyhd8ed1ab_0
zipp=3.7.0=pyhd8ed1ab_1
zlib=1.2.11=h7f8727e_4
zstd=1.4.9=haebb681_0
